+++ Cargo.toml
@@ -38,11 +38,8 @@ windows = { version = "0.52.0", features = [
 
 [target.'cfg(target_os = "linux")'.dependencies]
 xcb = "1.3.0"
-
-[target.'cfg(target_os = "linux")'.dependencies.x11]
-version = "2.21.0"
-optional = false
-features = ["xlib"]
+x11 = { version = "2.21.0", features = ["xlib"] }
+zbus = { version = "4.0.1" }
 
 [target.'cfg(target_os = "macos")'.dependencies]
 cocoa = "0.25.0"
