+++ src/linux/mod.rs
@@ -1,8 +1,8 @@
 #![deny(unused_imports)]
 
 mod api;
-
 use crate::common::api::API;
+
 use api::LinuxAPI;
 
 pub fn init_platform_api() -> impl API {
